relocated interface
